pyproject.toml
[tool.poetry]
name = "sections"
version = "0.1.0"
description = ""
authors = ["PrithviVK <p.lnu001@umb.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.7"
langgraph = "^0.1.6"
langchain-openai = "^0.1.14"
python-decouple = "^3.8"
duckduckgo-search = "^6.1.12"
gradio = "^4.37.2"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
